fixed the test based on the merge
